refactor due to comments [2]
