+++ packages/data/lib/repository/movie_repository.dart
@@ -1,8 +1,8 @@
-import 'package:data/service/api_base_service.dart';
+import 'package:data/service/api_service.dart';
 import 'package:data/service/service_payload.dart';
 import 'package:data/utils/constants.dart';
 import 'package:domain/di/data_model.dart';
-import 'package:domain/repository/base_repository.dart';
+import 'package:domain/repository/network_repository.dart';
 
 class NetworkRepositoryImpl implements NetworkRepository {
   final ApiService<ServicePayload> _apiService;
