+++ packages/domain/lib/model/data_model.dart
@@ -1,9 +1,9 @@
 class GetDataResponse {
-  final Map<String, dynamic> headers;
-  final List<dynamic> body;
-
   const GetDataResponse({
     required this.headers,
     required this.body,
   });
+
+  final Map<String, dynamic> headers;
+  final List<dynamic> body;
 }
