+++ packages/domain/lib/model/people_and_images_model.dart
@@ -1,11 +1,11 @@
 class PeopleAndImagesModel {
-  final String? characters;
-  final String? person;
-  final String? image;
-
   const PeopleAndImagesModel({
     required this.characters,
     required this.person,
     required this.image,
   });
+
+  final String? characters;
+  final String? person;
+  final String? image;
 }
