+++ packages/domain/lib/repository/network_repository.dart
