+++ packages/presentation/lib/config/theme/app_colors.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 
 abstract class AppColorsDark {
+  AppColorsDark._();
+
   static const primaryColorDark = Color.fromRGBO(15, 27, 43, 1);
   static const canvasColor = Color.fromRGBO(15, 27, 43, 1);
   static const borderTabBar = Color.fromRGBO(44, 63, 91, 1);
