+++ packages/presentation/lib/screen/movie_details/model/people_model.dart
@@ -1,11 +1,11 @@
 class PeopleModel {
-  final String image;
-  final String name;
-  final String playName;
-
   const PeopleModel({
     required this.image,
     required this.name,
     required this.playName,
   });
+
+  final String image;
+  final String name;
+  final String playName;
 }
